Move `GeniusYield.TxBuilder.Clb` -> `GeniusYield.Test.Clb`
