+++ src/GeniusYield/TxBuilder.hs
@@ -25,7 +25,6 @@ import           GeniusYield.TxBuilder.Errors    as X
 import           GeniusYield.TxBuilder.Node      as X
 import           GeniusYield.TxBuilder.NodeQuery as X
 import           GeniusYield.TxBuilder.Random    as X
-import           GeniusYield.TxBuilder.Clb       as X
 
 import           GeniusYield.Imports
 import           GeniusYield.Types
